requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
black==24.2.0
blinker==1.7.0
cachetools==5.3.2
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.4
dataclasses-json-speakeasy==0.5.11
distro==1.9.0
emoji==2.10.1
fastapi==0.109.2
filetype==1.2.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.42
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
idna==3.6
importlib-metadata==7.0.1
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
langchain==0.1.8
langchain-community==0.0.21
langchain-core==0.1.24
langdetect==1.0.9
langsmith==0.1.3
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.2
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.26.4
openai==1.12.0
packaging==23.2
pandas==2.2.0
pathspec==0.12.1
pillow==10.2.0
pinecone-client==3.0.3
platformdirs==4.2.0
protobuf==4.25.3
pyarrow==15.0.0
pydantic==2.6.1
pydantic_core==2.16.2
pydeck==0.8.1b0
Pygments==2.17.2
python-dateutil==2.8.2
python-iso639==2024.2.7
python-magic==0.4.27
python-magic-bin==0.4.14
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.6.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rpds-py==0.18.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.27
starlette==0.36.3
streamlit==1.31.1
streamlit-chat==0.1.1
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.6.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
unicorn==2.0.1.post1
unstructured==0.12.4
unstructured-client==0.18.0
urllib3==2.2.1
validators==0.22.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
